Move pseudo VVH timestamp creation from GeometryUtils to VVHClient
