+++ digiroad2-geo/src/main/scala/fi/liikennevirasto/digiroad2/GeometryUtils.scala
@@ -1,8 +1,5 @@
 package fi.liikennevirasto.digiroad2
 
-import org.joda.time.{DateTime, DateTimeZone}
-import org.joda.time.tz.DateTimeZoneBuilder
-
 object GeometryUtils {
   def geometryEndpoints(geometry: Seq[Point]): (Point, Point) = {
     val firstPoint: Point = geometry.head
@@ -213,19 +210,6 @@ object GeometryUtils {
 
   def isDirectionChangeProjection(projection: Projection): Boolean = {
     ((projection.oldEnd - projection.oldStart)*(projection.newEnd - projection.newStart)) < 0
-  }
-
-  /**
-    * Creates a pseudo VVH Timestamp for new assets and speed limits. Turns clock back to 0:00 on the same day
-    * if less than offsetHours have passed since or 0:00 on previous day if not.
-    *
-    * @param offsetHours Number of hours since midnight to return current day as a VVH timestamp (UNIX time in ms)
-    */
-  def createVVHTimeStamp(offsetHours: Int): Long = {
-    val oneHourInMs = 60 * 60 * 1000L
-    val utcTime = DateTime.now().minusHours(offsetHours).getMillis
-    val curr = utcTime + DateTimeZone.getDefault.getOffset(utcTime)
-    curr - (curr % (24L*oneHourInMs))
   }
 
   case class Projection(oldStart: Double, oldEnd: Double, newStart: Double, newEnd: Double, vvhTimeStamp: Long)
